requirements.txt
pandas>=2.0,<3
numpy>=1.24,<3
requests>=2.31
python-dotenv>=1.0
psycopg2-binary>=2.9,<3
pytest>=8.0
rapidfuzz>=3.0
nba_api>=1.5
